* 'develop' of https://github.com/CenterForOpenScience/osf.io: (304 commits)
  Use VOL-safe redirect
  Handle Session data async for PageCounters
  Prevent synchronous PageCounter locking
  Compose NODE_ADDON_R/W into NODE_DATA_R/W
  Update CHANGELOG, package.json
  Collections Moderation (#10131)
  fix flake8
  update tests
  construct a url for checking domains
  Prevent floating point domains
  really follow redirects
  Add NotableDomain perms to osf_admin
  Fix weird redirect case
  Bump version and update changelog.
  Feature/spam mitigation (#10127)
  Fix ES contributor serialization
  Switch OKState SSO from CAS to SAML
  Optimize filtering nonnull parents
  fix erroneous zeroes in PreprintSummaryReport
  Update OSFI nesta SSO: from okta to jumpcloud
  ...

# Conflicts:
#	api_tests/nodes/views/test_node_draft_registration_detail.py
#	api_tests/nodes/views/test_node_draft_registration_list.py

+++ website/search/elastic_search.py
@@ -1,4 +1,3 @@
-
 # -*- coding: utf-8 -*-
 
 from __future__ import division
@@ -32,6 +31,7 @@ from osf.models import SpamStatus
 from addons.wiki.models import WikiPage
 from osf.models import CollectionSubmission
 from osf.utils.sanitize import unescape_entities
+from osf.utils.workflows import CollectionSubmissionStates
 from website import settings
 from website.filters import profile_image_url
 from osf.models.licenses import serialize_node_license_record
@@ -402,11 +402,11 @@ def serialize_node(node, category):
         'id': node._id,
         'contributors': [
             {
-                'fullname': x['fullname'],
-                'url': '/{}/'.format(x['guids___id']) if x['is_active'] else None
+                'fullname': x['user__fullname'],
+                'url': '/{}/'.format(x['user__guids___id']) if x['user__is_active'] else None
             }
-            for x in node._contributors.filter(contributor__visible=True).order_by('contributor___order')
-            .values('fullname', 'guids___id', 'is_active')
+            for x in node.contributor_set.filter(visible=True).order_by('_order')
+            .values('user__fullname', 'user__guids___id', 'user__is_active')
         ],
         'groups': [
             {
@@ -454,11 +454,11 @@ def serialize_preprint(preprint, category):
         'id': preprint._id,
         'contributors': [
             {
-                'fullname': x['fullname'],
-                'url': '/{}/'.format(x['guids___id']) if x['is_active'] else None
+                'fullname': x['user__fullname'],
+                'url': '/{}/'.format(x['user__guids___id']) if x['user__is_active'] else None
             }
-            for x in preprint._contributors.filter(preprintcontributor__visible=True).order_by('preprintcontributor___order')
-            .values('fullname', 'guids___id', 'is_active')
+            for x in preprint.contributor_set.filter(visible=True).order_by('_order')
+            .values('user__fullname', 'user__guids___id', 'user__is_active')
         ],
         'title': preprint.title,
         'normalized_title': normalized_title,
@@ -584,48 +584,57 @@ def bulk_update_nodes(serialize, nodes, index=None, category=None):
     if actions:
         return helpers.bulk(client(), actions)
 
-def serialize_cgm_contributor(contrib):
+
+def serialize_collection_submission_contributor(contrib):
     return {
-        'fullname': contrib['fullname'],
-        'url': '/{}/'.format(contrib['guids___id']) if contrib['is_active'] else None
+        'fullname': contrib['user__fullname'],
+        'url': '/{}/'.format(contrib['user__guids___id']) if contrib['user__is_active'] else None
     }
 
-def serialize_cgm(cgm):
-    obj = cgm.guid.referent
+
+def serialize_collection_submission(collection_submission):
+    obj = collection_submission.guid.referent
     contributors = []
-    if hasattr(obj, '_contributors'):
-        contributors = obj._contributors.filter(contributor__visible=True).order_by('contributor___order').values('fullname', 'guids___id', 'is_active')
+    if hasattr(obj, 'contributor_set'):
+        contributors = obj.contributor_set.filter(visible=True).order_by('_order').values('user__fullname', 'user__guids___id', 'user__is_active')
+
+    tags = []
+    if hasattr(obj, 'tags'):
+        tags = list(obj.tags.filter(system=False).values_list('name', flat=True))
 
     return {
-        'id': cgm._id,
+        'id': collection_submission._id,
         'abstract': getattr(obj, 'description', ''),
-        'contributors': [serialize_cgm_contributor(contrib) for contrib in contributors],
-        'provider': getattr(cgm.collection.provider, '_id', None),
-        'modified': max(cgm.modified, obj.modified),
-        'collectedType': cgm.collected_type,
-        'status': cgm.status,
-        'volume': cgm.volume,
-        'issue': cgm.issue,
-        'programArea': cgm.program_area,
-        'subjects': list(cgm.subjects.values_list('text', flat=True)),
+        'contributors': [serialize_collection_submission_contributor(contrib) for contrib in contributors],
+        'provider': getattr(collection_submission.collection.provider, '_id', None),
+        'modified': max(collection_submission.modified, obj.modified),
+        'created': max(collection_submission.created, obj.created),
+        'collectedType': collection_submission.collected_type,
+        'status': collection_submission.status,
+        'volume': collection_submission.volume,
+        'issue': collection_submission.issue,
+        'programArea': collection_submission.program_area,
+        'schoolType': collection_submission.school_type,
+        'studyDesign': collection_submission.study_design,
+        'subjects': list(collection_submission.subjects.values_list('text', flat=True)),
         'title': getattr(obj, 'title', ''),
         'url': getattr(obj, 'url', ''),
-        'tags': list(obj.tags.filter(system=False).values_list('name', flat=True)),
+        'tags': tags,
         'category': 'collectionSubmission',
     }
 
 @requires_search
-def bulk_update_cgm(cgms, actions=None, op='update', index=None):
+def bulk_update_collection_submission(collection_submissions, actions=None, op='update', index=None):
     index = index or INDEX
-    if not actions and cgms:
+    if not actions and collection_submissions:
         actions = ({
             '_op_type': op,
             '_index': index,
-            '_id': cgm._id,
+            '_id': collection_submission._id,
             '_type': 'collectionSubmission',
-            'doc': serialize_cgm(cgm),
+            'doc': serialize_collection_submission(collection_submission),
             'doc_as_upsert': True,
-        } for cgm in cgms)
+        } for collection_submission in collection_submissions)
 
     try:
         helpers.bulk(client(), actions or [], refresh=True, raise_on_error=False)
@@ -667,14 +676,15 @@ def update_user(user, index=None):
             # update files in their quickfiles node if the user has been marked as spam
             if user.spam_status == SpamStatus.SPAM:
                 quickfiles = QuickFilesNode.objects.get_for_user(user)
-                for quickfile_id in quickfiles.files.values_list('_id', flat=True):
-                    client().delete(
-                        index=index,
-                        doc_type='file',
-                        id=quickfile_id,
-                        refresh=True,
-                        ignore=[404]
-                    )
+                if quickfiles:
+                    for quickfile_id in quickfiles.files.values_list('_id', flat=True):
+                        client().delete(
+                            index=index,
+                            doc_type='file',
+                            id=quickfile_id,
+                            refresh=True,
+                            ignore=[404]
+                        )
         except NotFoundError:
             pass
         return
@@ -794,7 +804,7 @@ def update_file(file_, index=None, delete=False):
 def update_institution(institution, index=None):
     index = index or INDEX
     id_ = institution._id
-    if institution.is_deleted:
+    if institution.deleted or institution.deactivated:
         client().delete(index=index, doc_type='institution', id=id_, refresh=True, ignore=[404])
     else:
         institution_doc = {
@@ -809,46 +819,42 @@ def update_institution(institution, index=None):
 
 
 @celery_app.task(bind=True, max_retries=5, default_retry_delay=60)
-def update_cgm_async(self, cgm_id, collection_id=None, op='update', index=None):
+def update_collection_submission_async(self, node_id, collection_id=None, op='update', index=None):
     CollectionSubmission = apps.get_model('osf.CollectionSubmission')
+    qs = CollectionSubmission.objects.filter(
+        guid___id=node_id,
+        collection__provider__isnull=False,
+        collection__deleted__isnull=True,
+        collection__is_bookmark_collection=False,
+    )
     if collection_id:
+        qs = qs.filter(collection_id=collection_id)
+
+    for collection_submission in qs:
+        if op == 'update':
+            if not collection_submission.guid.referent.is_public:
+                continue
+            if collection_submission.guid.referent.deleted:
+                continue
+            if collection_submission.state in [
+                CollectionSubmissionStates.REMOVED,
+                CollectionSubmissionStates.REJECTED,
+                CollectionSubmissionStates.PENDING
+            ]:
+                continue
         try:
-            cgm = CollectionSubmission.objects.get(
-                guid___id=cgm_id,
-                collection_id=collection_id,
-                collection__provider__isnull=False,
-                collection__deleted__isnull=True,
-                collection__is_bookmark_collection=False)
-
-        except CollectionSubmission.DoesNotExist:
-            logger.exception('Could not find object <_id {}> in a collection <_id {}>'.format(cgm_id, collection_id))
-        else:
-            if cgm and hasattr(cgm.guid.referent, 'is_public') and cgm.guid.referent.is_public:
-                try:
-                    update_cgm(cgm, op=op, index=index)
-                except Exception as exc:
-                    self.retry(exc=exc)
-    else:
-        cgms = CollectionSubmission.objects.filter(
-            guid___id=cgm_id,
-            collection__provider__isnull=False,
-            collection__deleted__isnull=True,
-            collection__is_bookmark_collection=False)
-
-        for cgm in cgms:
-            try:
-                update_cgm(cgm, op=op, index=index)
-            except Exception as exc:
-                self.retry(exc=exc)
+            update_collection_submission(collection_submission, op=op, index=index)
+        except Exception as exc:
+            self.retry(exc=exc)
 
 @requires_search
-def update_cgm(cgm, op='update', index=None):
+def update_collection_submission(collection_submission, op='update', index=None):
     index = index or INDEX
     if op == 'delete':
-        client().delete(index=index, doc_type='collectionSubmission', id=cgm._id, refresh=True, ignore=[404])
+        client().delete(index=index, doc_type='collectionSubmission', id=collection_submission._id, refresh=True, ignore=[404])
         return
-    collection_submission_doc = serialize_cgm(cgm)
-    client().index(index=index, doc_type='collectionSubmission', body=collection_submission_doc, id=cgm._id, refresh=True)
+    collection_submission_doc = serialize_collection_submission(collection_submission)
+    client().index(index=index, doc_type='collectionSubmission', body=collection_submission_doc, id=collection_submission._id, refresh=True)
 
 @requires_search
 def delete_all():
@@ -883,7 +889,9 @@ def create_index(index=None):
                     'programArea': NOT_ANALYZED_PROPERTY,
                     'provider': NOT_ANALYZED_PROPERTY,
                     'title': ENGLISH_ANALYZER_PROPERTY,
-                    'abstract': ENGLISH_ANALYZER_PROPERTY
+                    'abstract': ENGLISH_ANALYZER_PROPERTY,
+                    'schoolType': NOT_ANALYZED_PROPERTY,
+                    'studyDesign': NOT_ANALYZED_PROPERTY,
                 }
             }
         else:
